Got attack speed working.
Made attackType and baseBullet unrelated so that attackType is always instantiated as long as the player has that attack type on them. This way the default attack speed can be put on the player without a bullet being instantiated.

Bullets will get applied all of bonuses that bullets should have. Players though hold all of the bonsuses and attack types permanently